--- src/main/java/org/runimo/runimo/user/controller/IncubatingEggQueryUsecase.java
@@ -1,5 +1,0 @@
-package org.runimo.runimo.user.controller;
-
-public interface IncubatingEggQueryUsecase {
-  QueryIncubatingEggResponse execute(Long userId);
-}
