+++ src/main/java/org/runimo/runimo/user/service/usecases/eggs/GiveLovePointToEggUsecase.java
@@ -1,6 +1,6 @@
 package org.runimo.runimo.user.service.usecases.eggs;
 
-import org.runimo.runimo.user.controller.UseLovePointResponse;
+import org.runimo.runimo.user.service.dtos.UseLovePointResponse;
 import org.runimo.runimo.user.service.dtos.UseLovePointCommand;
 
 public interface GiveLovePointToEggUsecase {
