+++ src/main/java/org/runimo/runimo/user/service/usecases/eggs/IncubatingEggQueryUsecase.java
@@ -1,0 +1,7 @@
+package org.runimo.runimo.user.service.usecases.eggs;
+
+import org.runimo.runimo.user.service.dtos.QueryIncubatingEggResponse;
+
+public interface IncubatingEggQueryUsecase {
+  QueryIncubatingEggResponse execute(Long userId);
+}
